refactor(Legend): +PathsDates, +RatioTrends
