+++ src/components/Legend.js
@@ -2,179 +2,23 @@ import React from 'react';
 import differenceInDays from 'date-fns/differenceInDays';
 
 import './Legend.css';
-import Municipalities from './List/Combined/CITIES_SOCIAL_FRIENDLY/Municipalities';
-import { Arrow, Emoji } from './shared/ui/New';
 import apiPathObject from '../utils/apiPathObject';
 import { getDate, formatToLocaleDateString } from '../utils/dates';
+import RatioTrends from './Legend/RatioTrends';
+import PathsDates from './Legend/PathsDates';
 
-const municipalitiesTrend = [
-  {
-    description: 'Trend potrjenih primerov v občini pada.',
-    icon: { symbol: '⤵', attr: { role: 'img', ariaLabel: 'down' } },
-  },
-  {
-    description:
-      'Ni sprememb v trendu potrjenih primerov (trend v območju -0,03 do +0,03).',
-    icon: { symbol: '➖', attr: { role: 'img', ariaLabel: 'neutral' } },
-  },
-  {
-    description: 'Trend potrjenih primerov v občini raste.',
-    icon: { symbol: '⤴', attr: { role: 'img', ariaLabel: 'up' } },
-  },
-  {
-    description:
-      'Trenda ni mogoče izračunati (ena od vrednosti y1, y2, y3 je enaka 0).',
-    icon: {
-      symbol: 'brez',
-      attr: { role: 'img', ariaLabel: 'no symbol' },
-    },
-  },
-];
-
-function Legend({
-  municipalities,
-  css = {},
-  dates = {},
-  paths = {},
-  refreshData = {},
-}) {
-  const LegendSection = ({ title, children }) => (
-    <section className="legend-section">
-      <h3 className="bold legend-section-title">{title}:</h3>
-      {children}
-    </section>
-  );
-
-  const LegendTable = ({ data = [{}] }) => {
-    const tableBody = data.map(({ description, icon }, i) => {
-      const { symbol, attr } = icon;
-      const { role, ariaLabel } = attr;
-      return (
-        <tr key={`${i}-${ariaLabel}`}>
-          <td className="table-symbol">
-            <span role={role} aria-label={ariaLabel}>
-              {symbol}
-            </span>
-          </td>
-          <td className="table-description">{description}</td>
-        </tr>
-      );
-    });
-
-    return (
-      <table>
-        <thead>
-          <tr>
-            <th>Simbol</th>
-            <th>Opis</th>
-          </tr>
-        </thead>
-        <tbody>{tableBody}</tbody>
-      </table>
-    );
-  };
-
-  const apiDates = Object.entries(refreshData).map(([key, value], index) => {
-    if (key === 'today') {
-      return '';
-    }
-    return (
-      <li key={`${key}-${index}`} className={value[0]}>
-        <span className="bold">
-          {value[2].pathname}
-          {key === 'summary' ? '.casesActive' : ''}:
-        </span>{' '}
-        {value[1]}
-      </li>
-    );
-  });
-
-  const PathDate = ({ title, path }) => {
-    const markOk = <Emoji emoji={'✅'} ariaLabel="check mark" />;
-    const markFail = <Emoji emoji={'❌'} ariaLabel="cross mark" />;
-    // css[path] = '' || 'red'
-    return (
-      <li className={css[path]}>
-        {css[path] ? markFail : markOk} <span className="bold">{title}: </span>
-        <span className={css[path] && 'bold'}>{paths[path].pathname}</span>{' '}
-        <Arrow />{' '}
-        <span className={css[path] && 'bold'}>{dates[path].toString()}</span>
-      </li>
-    );
-  };
-
-  /**
-   * TODO split into two components <RefreshData/> & <RatioTrends/>
-   * RefereshData needs paths: stats, municipalities, patients, lab-tests, summary
-   * RatioTrends needs path: municipalities
-   */
+function Legend({ pathsDatesData, municipalitiesHook }) {
   // TODO move inline style to css file, when done with html
   return (
     <div className="Legend">
       <h2 id="legenda">Legenda</h2>
-      <LegendSection title={'Osveženi podatki'}>
-        <ul style={{ listStyle: 'none' }}>
-          <li>
-            <span className="bold">Danes: {dates.today.toString()}</span>
-          </li>
-          <li>
-            <span className="bold">API datumi:</span>
-            <ol style={{ padding: '8px 24px' }}>{apiDates}</ol>
-          </li>
-          <li>
-            {/* <TESTS_ACTIVE/> */}
-            <span className="bold">Testi in aktivni primeri:</span>
-            <ol style={{ padding: '8px 24px' }}>
-              <PathDate title={'PCR'} path={'labTests'} />
-              <PathDate title={'HAT'} path={'labTests'} />
-              <PathDate title={'Aktivni primeri'} path={'summary'} />
-            </ol>
-          </li>
-          <li>
-            {/* <HOSPITALIZED_DECEASED/> */}
-            <span className="bold">
-              Hospitalizirani, negovani in preminuli:
-            </span>{' '}
-            <ol className={css.patients} style={{ padding: '8px 24px' }}>
-              <PathDate title={'Hospitalizirani'} path={'patients'} />
-              <PathDate title={'Na respiratorju'} path={'patients'} />
-              <PathDate title={'Negovalne bolnišnice'} path={'patients'} />
-              <PathDate title={'Preminuli'} path={'patients'} />
-            </ol>
-          </li>
-          <li>
-            <span className="bold">Kombiniran:</span>
-            <ol style={{ padding: '8px 24px' }}>
-              {/* <Vaccination/> */}
-              <PathDate title={'Cepljeni'} path={'stats'} />
-              {/* <Confirmed/> */}
-              <PathDate title={'Potrjeni primeri'} path={'stats'} />
-              {/* <PerAge/> */}
-              <PathDate title={'Po starosti'} path={'stats'} />
-              {/* <InHospitals/> */}
-              <PathDate title={'Po bonišnicah'} path={'patients'} />
-              {/* <CITIES_SOCIAL_FRIENDLY/> */}
-              <PathDate title={'Po krajih'} path={'municipalities'} />
-            </ol>
-          </li>
-        </ul>
-      </LegendSection>
-      <LegendSection
-        title={'Trend rasti potrjenih primerov v posamezni občini'}
-      >
-        <LegendTable data={municipalitiesTrend} />
-      </LegendSection>
-      <LegendSection title={'Formula za izračun trenda'}>
-        <p>trend = ( log(y1) + 3 * log(y3) - 4*log(y2) ) / 8</p>
-        <p>y1 = vsota novih primerov za dneve (-14..-8)</p>
-        <p>y2 = vsota novih primerov za dneve (-10..-4)</p>
-        <p>y3 = vsota novih primerov za dneve (-6..0)</p>
-      </LegendSection>
-      <LegendSection title={'Občine CHECK ratio'}>
-        <ul className="municipalities">
-          <Municipalities data={municipalities} showTrend="n"></Municipalities>
-        </ul>
-      </LegendSection>
+      <PathsDates
+        css={pathsDatesData?.css}
+        dates={pathsDatesData?.dates}
+        paths={pathsDatesData?.paths}
+        refreshData={pathsDatesData?.refreshData}
+      />
+      <RatioTrends municipalitiesHook={municipalitiesHook} />
     </div>
   );
 }
@@ -205,20 +49,22 @@ function withLegendHOC(Component) {
       summary !== null &&
       hospitalsList !== null;
 
-    const legendProps =
+    const pathsDateData =
       allDataExists &&
-      getLegendData({ stats, patients, municipalities, labTests, summary });
+      getPathsDatesData({ stats, patients, municipalities, labTests, summary });
 
-    if (legendProps) {
-      return <Component {...legendProps} />;
-    }
-    return null;
+    const legendProps = {
+      pathsDateData,
+      municipalitiesHook: props.municipalitiesHook,
+    };
+
+    return <Component {...legendProps} />;
   };
 }
 
 export default withLegendHOC(Legend);
 
-function getLegendData(
+function getPathsDatesData(
   allData = {
     stats: [],
     patients: [],
@@ -247,6 +93,7 @@ function getLegendData(
     summary: apiPathObject.summaryPath,
   };
 
+  // TODO find better name
   const refreshData =
     dates & css &&
     paths &&
